Don't strip p on display, breaks nls; use CSS to nullify p wrappers. ...
Markdown uses <p> tags to indicate paragraph breaks, so strip_p_tags
makes it so line breaks are not drawn. We wanted to strip_p_tags when we
were using lxml and getting an automatic p wrapper, but now we just want
to stop the extra space added at the beginning and end of each p-wrapped
block, so use CSS rules to nullify these in both comments and story
descriptions.