atom: Revert "Split up atom_combined.mak. ..."
This reverts commit 6f9ba0ec291602ee415d682605e06c0e69f73811.

We are reverting this because it appears that passing the SA object as
an argument between files in this way causes the SA session to detach,
and breaks the backreferences. If someone is bored, they should figure
out how to get around that so we can implement this fix again. @FIXME.