section: Use uploaded logo as-is. ...
This could present a security problem at some point. Re-encoding the
image ensures that we use a version without any nefarious or invisible
code in place, but we are disabling it for now as this is currently an
admin-only feature and our pictures are happier if they are not run
through the ringer needlessly.