notify: Pretend to succeed if user has no email. ...
This works OK now because we aren't really tracking these mails, but
will have to change if/when we start. For now, it makes things smoother,
and notifications will work fine as soon as the user adds an email.

This also does a lot of unnecessary work just to get thrown away, so we
can/should place the exit up earlier as a performance improvement. We
are doing this in its current location now because the function already
had the data we needed.